requirements.txt
# Scratch × Gemini AI サーバー 依存ライブラリ
flask>=3.0.0
flask-cors>=4.0.0
supabase>=2.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
